feat: Implement dynamic status background and text color
This commit introduces the following changes:

Modified the admin-app to display a spinner for class status, allowing for easy updates.

Implemented dynamic background and text color changes for the status spinner based on the selected status (Scheduled, Completed, Cancelled).

Updated the user-app to display the class status with the corresponding background color on both the class list and class detail screens.

Recreated and deployed the resetDatabase Firebase Function with a secret key for authentication and upgraded to Node.js 20.